--- apps/diffusion/src/search/CardMap.css
@@ -1,87 +1,0 @@
-.mapboxgl-popup-content {
-  padding: 0px;
-  border-radius: 5px;
-  box-shadow: 0 2px 4px 0 rgba(189, 189, 189, 0.5);
-  transition: 0.3s;
-}
-
-.mapboxgl-popup-content:hover {
-  box-shadow: 0 3px 6px 0 rgba(189, 189, 189, 1);
-}
-
-.map-card {
-  display: flex;
-  flex-direction: row;
-  overflow: hidden;
-  background-color: white;
-  height: 150px;
-  width: 800px;
-}
-
-.map-card .content {
-  padding: 0px 10px 0px 10px;
-  overflow: hidden;
-  width: 100%;
-}
-.map-card .thumbnail {
-  min-width: 120px;
-  max-width: 150px;
-  max-height: 150px;
-  background-color: #f8f8f8;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.map-card img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-
-.map-card img.no-img {
-  width: 190px;
-  height: 140px;
-}
-
-.map-card .content .categories {
-  font-size: 9px;
-  color: #808d9e;
-  font-weight: 300;
-  margin-bottom: 10px;
-}
-
-.map-card h2 {
-  font-weight: 600;
-  font-size: 22px;
-  color: #2a282b;
-  margin-bottom: 0;
-}
-
-.map-card h3 {
-  font-weight: 600;
-  font-size: 16px;
-  color: #2a282b;
-  margin-bottom: 0;
-}
-
-.map-card span {
-  font-weight: 400;
-  font-size: 12px;
-  color: #19414c;
-  margin-left: auto;
-}
-
-.map-card p {
-  color: #2a282b;
-  font-weight: 400;
-  font-size: 14px;
-  margin-bottom: 5px;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.map-card .base {
-  float: right;
-}
