+++ apps/diffusion/src/search/MapList/LinkedNotices.css
