+++ apps/diffusion/src/search/MapList/Map.css
