+++ apps/diffusion/src/search/MapList/SingleNotice.css
