+++ apps/diffusion/src/search/MapList/mapbox-gl.css
