+++ apps/diffusion/src/search/MosaicList/CardMosaic.css
