+++ apps/diffusion/src/search/MosaicList/CardMosaic.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Link from "next/link";
 import { Col } from "reactstrap";
-import { image } from "./image";
+import { image } from "../utils/image";
 import "./CardMosaic.css";
 export default ({ data }) => (
   <Col>
