+++ apps/diffusion/src/search/MosaicList/Mosaic.js
