+++ apps/shared/package.json
@@ -1,6 +1,6 @@
 {
   "name": "pop-shared",
-  "version": "0.9.0",
+  "version": "0.9.1",
   "description": "pop shared components",
   "author": "betagouv",
   "license": "MIT",
